requirements.txt
ConcurrentLogHandler==0.9.1
Flask-RESTful==0.3.6
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
httplib2 == 0.9.1
jsonschema==2.6.0
oslo.config == 1.4.0
ping==0.2
sqlalchemy == 1.0.12
requests == 2.20.1
sphinx == 1.8.2
sphinx-rtd-theme == 0.4.2
flask-login==0.4.1
flask-migrate==2.3.1
gunicorn==19.4.5
eventlet==0.24.1
